Bug fix: Update HISTORY.rc templates for latest State_Chm fields
The HISTORY.rc template files were never updated to use the latest
StateChm collection names (e.g. pHaccumMode, phCoarseMode, HPlusAccumMode,
HPlusCoarseMode, etc.)  This is now fixed.

Also, for now we have commented out the CloudConvFlux, WetLossConv,
and WetLossLS collections for the GCHP template file.  These collections
are not needed for the benchmark, and causes GCHP benchmark simulations
to die with a "file too long" input error.

NOTE: Further modifications to the run directory script may be needed,
but these will be done in a subsequent commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>