Fix GCHP bug: benchmark sim name change omitted some vars in restart
GCHP benchmark requires certain variables be present in the restart file,
including area, box height, and trop level. Previously these were added
to the internal state based on simulation name in input.geos. However,
the simulation name in input.geos changed from benchmark to fullchem,
but the string update in the conditional to add the internal vars
remained benchmark.

With this update the extra variables will be added to all fullchem
simulation restarts so that unit conversions with the output variables
is possible. They were already added to the transport tracer simulation
restart.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>