Add initial modifications for scaling carbon units explicitly in HEMCO_Config.rc
New scale factors have been added to HEMCO_Config.rc to convert from units
carbon to units species. These have been applied to NEI2011 entries so far,
but still need to be applied to remaining entries where input units are
in kgC.

The scale factors have been reordered in HEMCO_Config.rc so that they are
in increasing order as much as possible. This is to help identify what
scale factor values are already in use.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>