Replace call to Convert_Spc_Units for species diags with local conversion
Calling Convert_Spc_Units to convert to and from units introduces small
numerical differences in State_Chm%Species. It therefore should only be
used if necessary, such as for changing units for use in a GEOS-Chem
core component such as chemistry. However, it was also used to convert
units for assigning several State_Diag arrays for species diagnostics,
introducing small but unnecessary differences into the model.

This is now fixed by replacing the calls to Convert_Spc_Units in
diagnostics_mod.F90 with use of a local array equal to State_Chm%Species
but with units converted to v/v dry.

Also in this commit, outdated comments about units in main.F are removed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>