Rename State_Chm variables with suffix SAV; delete unused var AcidPurSav
Several State_Chm variables are renamed for clarity:
1. DryDepSav -> DryDepFreq
2. pHSav     -> IsorropAeropH
3. HplusSav  -> IsorropHplus
4. WaterSav  -> IsorropAeroH2O
5. SulratSav -> IsorropSulfate
6. NaRatSav  -> IsorropNitrate
7. BisulSav  -> IsorropBisulfate

Chemistry state ID used for registering diagnostics also changed to better
match the new State_Chm array names. All relevant HISTORY.rc files are
updated with the new names.

Long name description for IsorropNitrate updated to reflect it is nitrate
and not sodium.

Existing variable State_Chm%AcidPurSav is deleted. It is set in Isorropia
but is never used for anything.

Subroutine Update_DryDepSav is renamed to Update_DryDepFreq to reflect
the new naming.

All local variable instances of DEPSAV are renamed to be DepFreq.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>