Replace runScriptSamples directory with symlink to source in GCHP rundir
The runScriptSamples directory is stored in the source code and no
modifications are made when copying to the run directory. It is therefore
more efficient to point to the directory in the source code rather than
copy all the example files.

This update also removes a residual symbolic link created for gFTL in
GCHP run directories that is no longer needed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>