Update HEMCO configuration files to use CEDS by default, list years available, and use RF flag for regional emissions
Daniel Jacob wrote to the GEOS-Chem Steering Committee:

   Here’s what I suggest as take-away:

   1. We adopt CEDS as our global default inventory in version 13.0.0 and
      use it for benchmarking with no regional overwrites for now, as
      representing our best estimate of emissions for use in global model
      applications.  This is the prevailing GCSC view. CEDS already subsumes
      all the regional emission inventories currently in HEMCO and has the
      advantage of temporal consistency out to 2017.  The regional
      inventories have higher spatial resolution and better speciation of
      VOC emissions, but these are of most interest for nested model users
      of which they are many but I expect that these nested model users pay
      particular attention to the emissions they use and may therefore
      choose to use the regional inventories.
   2.  As to Seb’s good point of exercising a maximum of model options in
       the benchmark, we do embrace this with OA as you know – but that’s
       because we don’t have a recommendation on OA for global model users,
       and because the different OA options could well break with model
       updates.  We don’t do this for any of the other model options. I see
       no reason that the regional inventories would break with model updates.
   3. We will supersede CEDS with new regional emission inventories in the
      default in the future if we decide that they are objectively better.
      For example, we expect to have soon an updated NEI for US emissions
      from Barron and I would recommend to have it overwrite CEDS in the
      default.
   4. We should still overwrite CEDS with the Tzompa ethane and Xiao
      propane because we have as of now no assurance that CEDS does better
      for those. It would be great for someone to look into it!

The GEOS-Chem Steering Committee also expressed concern about users unknowingly
running with emissions for past years due to the use of HEMCO's 'C' time
cycle flag indicating the latest emissions year available should be used.
To resolve this, we have changed the time cycle flag to 'RF' for regional
inventories that are provided for multiple years. This will force HEMCO
to crash with an error that the current time isn't available if a user
attempts to use the inventory for a year outside of it's coverage. Users
can always manually revert to the 'C' time cycle flag to use the latest year
available. We do not use the 'RF' flags for global inventories (e.g. CEDS)
because this would cause all GEOS-Chem and HEMCO standalone simulations
for 2018 and after to crash due to unavailable data. The benchmark simulations
are now for 2019 and would crash in that case, so we keep the time cycle
flag for CEDS and other global inventories as 'C'.

Finally, comments have been added to the emission switches section to make it
clear what years each emissions inventory and dataset are available. This
should help users when choosing which inventories to turn on/off for their
simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>