Bug fix: bypass MAPL read issue for lightning by using daily files
Previously GCHP (and GEOS-Chem Classic still) read offline lightning files
that are stored in monthly files. This causes an issue in GCHP for some
months due to logic in the way MAPL handles file time calculation.
Switching to daily files corrects the issue. The lightning data itself is
not different.

Note that the paths in HEMCO_Config.rc are updated for consistency with
ExtData.rc eventhough they are not used.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>