Update GCClassic integration test scripts to compile only a few times
GEOS-Chem integration tests now only use 5 compilation tests
(default, with APM, with BPCH diags, with RRTMG, with TOMAS). This
results in much faster execution tests.  We have updated the scripts
in test/GCClassic accordingly.

NOTE: I had intended to store the executable files in the $root/exe_files
folder but this does not seem to work properly.  This will be resolved
in a future commit.  For now the executables are placed in the
$root/build folder.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>