Bug fix: remove SOAP and SOAS from input.geos for complexSOA* simulations
We have added an extra check in run/shared/setupConfigFiles.sh
to remove the SOAP and SOAS lines from input.geos.  Having these
present causes a seg fault error due to the fact that it will
try to execute the "IF ( SimpleSOA )" blocks in aerosol_mod.F90.

NOTE: This does not affect the benchmark simulation, since that
needs to have SOAP and SOAS defined.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>