Add AgeOfAir to internal state and now update from GCHP_Chunk_Run
AgeOfAir [s] is a field in State_Met used to diagnose transport. It will be
used in the 10-year stratospheric benchmarks. In GEOS-Chem Classic, AgeOfAir
will be saved to the restart file through HISTORY.rc, but for GCHP we need
to add it to MAPL's internal state. Here a call to update AgeOfAir on each
transport is also added to gchp_chunk_mod.F90.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>