Use a better algorithm in function absolute_path
For function absolute_path in commonFunctionsForTests.sh, we now
use a better algorithm.  If the path exists, we use readlink -f
to get the absolute path.  Otherwise we just replace ~ with $HOME.
This avoids issues with paths that don't exist.

In the intTestCreate.sh scripts, we make sure to call absolute_path
for the root directory again after the root folder is made.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>