Add fix to save monthly diagnostic output in GCHP multi-segmented runs
There is code in gchp.multirun.run to handle setting the number of hours per
month and replacing the frequency and duration for diagnostics with that
value in runConfig.rc (and eventually HISTORY.rc). This code needed to
be updated to replace the values for timeAvg_freq, timeAvg_dur, inst_freq,
and inst_dur instead of common_freq and common_dur which do not exist in
runConfig.sh anymore.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>