Add parallel KPP Arrhenius functions GCARR1, GCARR2
In the gckpp.kpp and fullchem.eqn files, many reaction rates were defined
with the GCARR (Arrhenius law) which computes the expression:

 a0 * EXP( c0 / T ) * ( 300 / T )**b0

given the temperature T and 3 Arrhenius coefficients a0, b0, c0.

However, it is more computationally efficient to return

 a0 * EXP( c0 / T )

if b0 is zero (since the expression using b0 will evaluate to 1)

and likewise, it is more computationally efficient to return

 a0 * ( 300 / T )**b0

if c0 is zero.

Therefore, we have defined 2 additional functions in gckpp.kpp:
 - GCARR1 (to be used when c0 = 0)
 - GCARR2 (to be used when b0 = 0)

Also note, if both b0 = 0 and c0 = 0, we don't need to call a function,
as we can just specify the value a0 as the reaction rate.

We have adjusted the RHS of reaction rate definitions in fullchem.eqn
accordingly.  These updates have produced identical results w/r/t the
reference code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>