Fixes for minor bugs that prevented TOMAS from building & running
GeosCore/gamap_mod.F90: Fixed a formatting error in routine WRITE_TINFO.
When the moles C/moles species value was removed, this caused a mismatch
in the format statement.  We now write "1" as a placeholder, which
eliminates the formatting error.

run/shared/setupConfigFiles.sh: The line "Use non-local PBL" ws changed
to " => Use non-local PBL", so the "replace_colon_sep_val" search &
replace function was missing this.  We have now fixed it so that
replace_colon_sep_val now looks for the proper text.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>