Bug fix: Set METEOROLOGY switch to false for TransportTracers HEMCO template
The METEOROLOGY main switch was inadvertently in the
runs/GCHP/HEMCO_Config.rc.templates/HEMCO_Config.rc.TransportTracers file
to true (probably a cut-n-paste error).  This needs to be set to
false, because ExtData reads the meteorology.

This bug caused a bunch of GCHP errors such as:
 ERROR: Timer TOTAL needs to be set first
 ERROR: Timer INITIALIZE needs to be set first
that cause the run to crash.

This is now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>