Add fixes to prevent div-by-zero in gckpp_HetRates.F90
We now have rewritten the logic when computing the brConc_Cld* and
clConc_Cld* variables in order to avoid division-by-zero errors.

Also, we have noticed that the denominator in these expressions has
was being computed over and over again.  We now compute these once
and reuse them, which is more efficient.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>