Recompute denominators after GET_HALIDE_CLDCONC in gckpp_HetRates.F90
In the prior commit (75b4ba7), the denominators used to compute the
CldConc_Brg* and CldConc_Clg* quantities were computed before the
call to GET_HALIDE_CONC.  This was incorrect, because it introduced
some numerical differences into the output.

The simple solution is to recompute the denominators after the
call to GET_HALIDE_CLDCONC.  This resulted in identical output
w/r/t the prior commit (75b4ba7).

Also trimmed trailing whitespace.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>