Rename rate law functions; update GCJPLPR and GCLJPLEQ functions
We now use the nomenclature _abc to denote the Arrhenius parameters that
are passed to each function.  For example GCARR_abc takes all 3 parameters
A, B, and C, but GCARR_ac only takes A & C (since B=0 and that would
result in useless CPU cycles to evaluate a term that goes to 1.)

Also split the GCJPLPR function into multiple functions, depending on
which Arrhenius factors are nonzero.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>