Replace hard-coded resolution strings in GCHP's ExtData.rc with tokens
The offline emissions entries in ExtData.rc were using hard-coded resolutions
(0.5x0.625) in the file path. This would cause problems if using GEOS-FP
met fields, which should use the files with resolution 0.25x0.3125. To fix
this, we now use tokens in ExtData.rc which are automatically replaced with
the appropriate native resolution string by the createRunDir.sh script.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>