More stable stratospheric H2O boundary condition
By selecting the "static" H2O option in input.geos, the user
can now request that stratospheric water vapor be specified
up to 70 hPa in the tropics. This results in a more stable
water vapor signal in the stratosphere. If this option is
not selected, the water vapor in the stratosphere may grow
continuously over multi-year simulations.