Change time cycle flag for species concentrations from restart files to force date to match
The time cycle flags for the species concentration fields (SPC_) from the
restart files were previously either (EY or CYS). In the case of "EY",
HEMCO would only read the restart file if the date in the restart file
matched the simulation date, but would not crash if the correct date was
not found. If GEOS-Chem then couldn't find the field from HEMCO, it would
set initial species concentrations to the default background values specified
in the species database. In the case of "CYS", HEMCO would use the restart
file for the closest available date. If it couldn't find a restart file then
HEMCO would skip that field and GEOS-Chem would default to the background
concentrations. In both cases, the user might not be aware of this behavior
unless they inspected the GEOS-Chem log file.

To avoid using restart files for the wrong date and/or defaulting to
background species concentrations without the user's knowledge, we now use
the "EFY" flag for the SPC_ fields. This tells HEMCO to read the initial
species concentrations from the restart file only if the date in the file
matches the simulation date. If a restart file is not found, then the
simulation will crash. Users can then take it upon themselves to modify the
time cycle flag, but they will at least do so knowing how the initial
concentrations will be handled by HEMCO.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>