Add fixes for restart files during GEOS-Chem Classic run directory creation
Set start date for  only Hg, CH4, and TransportTracer simulations to
20190101. All other simulations use a start date of 20190701 since the
sample restart files for those simulations are for July.

The restart files in GEOSCHEM_RESTARTS/v2020-02 have been fixes so that
the names are in the format GEOSChem.Restart.{sim_name}.YYYYMMDD_0000z.nc4.
This allows users to easily see what date the restart files represent and
is more consisteny with the naming format in the run directories. The
timestamps in those files have also been fixed so that hour is 0000z (some
files previously had erroneous hhmm values.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>