Allow for missing species in restart files for certain simulations
The sample restart file provided for some simulations does not include all
species for that simulation. For example, the fullchem restart file has been
saved out for the benchmark simulation and does not include aciduptake,
marinePOA, complexSOA_SVPOA, or APM species. For these simulations, we need
to change the time cycle flag in HEMCO_Config.rc from "EFY" to "EY" to allow
HEMCO to skip the missing species which will result in GEOS-Chem assigning
default background concentrations for those species. A warning will be
printed to the screen when a user generates a run directory for one of these
simulations letting them know that some species may be assigned background
concentrations for their initial concentrations.

This commit also fixes a typo in the TOMAS restart file names.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>