Met field entries in HEMCO_Config.rc template files now use flag EFYK
This commit (and also commit geoschem/HEMCO 2e4df52) addresses a problem
that arose in GEOS-Chem 13.0.1. In that version, the time cycle flag
"EFY" was changed to "exact, force error if not found, use simulation
year, read only once".  This was done in order to make sure that the
restart file date matched the user's simulation start date; otherwise
an error is thrown.

However, the met field entries in HEMCO_Config.rc also used this EFY
time cycle flag.  As a result, the modification of behavior to EFY
meant that met fields were no longer being read from disk as the
situation evolved.

To fix this situation, we have created a new HEMCO time cycle flag,
called "EFYK".  This is similar to "EFY", but will cause the data to
be continuously read during a simulation.

In this commit we have updated the HEMCO_Config.rc entries in the
run/GCClassic and run/GCHP directories accordingly.  (NOTE: This will
only affect GCClassic runs as GCHP uses MAPL to read data.  But we have
added EFYK flags to GCHP HEMCO_Config.rc templates for consistency.)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>