Add fixes for GCHP run directory creation
The debug options were removed from runConfig.sh.template in an earlier
commit, but the replace_val lines for the debug options remained. This
caused the script to hang when creating a GCHP run directory. The solution
is to remove the replace_val lines for DEBUG_LEVEL and MEMORY_DEBUG_LEVEL.

The runCondig_adj.sh.template file has also been updated here so that
it is consistent with recent updates to runConfig.sh.template for the
forward model.