Add earlier warning in planeflight_mod.F90 for surface observations
There was a warning in planeflight_mod.F90 when computing fields for the
planeflight diagnostic to tell users that the observation is being considered
a surface observation and thus level is forced to L=1. This warning has
been removed and instead a warning now prints when planeflight observations
are read from file. If an observation type name begins with 'S' the code will
stop letting the user know that the code automatically assumes this is a
surface observation. If that's not intended, the user may disable the code
forcing the unit change and L=1 or change the observation type name.

This fix addresses Github issue https://github.com/geoschem/geos-chem/issues/680.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>