Fix array-out-of-bounds error in dust_mod.F90 for APM microphysics sims
In GeosCore/dust_mod.F90, there was code that computed IDDST as the
HEMCO species Id of DST1, and then assumed that the other dust
species were sequentially following.  This seems to have caused an
out-of-bounds error.

We now save the HEMCO DST1-DST4 species ID's in a vector that is used
to get emissions from the HEMCO state.  Furthermore, we have put an
error trap that will cycle in case a dust emissions is missing (and
its HEMCO state array is unallocated).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>