Fix how diurnal scale factors of Chinese power plants are applied in HEMCO_Config.rc
The original implementation of the diurnal scale factors of Chinese power
plants were applied to CEDS emissions by adding *_ENE_CH to HEMCO_Config.rc
with hierarchy increased by 1 to overwrite those emissions over China.
This led to the other CEDS emission sectors getting zeroed out, since they
still had the original hierarchy. To fix this, we now apply scale factors to
(1) undo the EDGAR and GEIA diurnal scale factors over China (using oper=-1
to divide and mask=1009 for China) and (2) apply the PKU scale factors using
mask=1009 to only scale over China.

Also in this commit, the HEMCO_Config.rc for GCHP full-chemistry simulations
is updated to be consistent with the GCClassic version. Mainly, this involved
bringing in this new scaling and new emission inventories introduced by the
GCAP 2.0 updates. Note: Entries for these inventories have not been added
to ExtData.rc in GCHP, so if they are turned on in HEMCO_Config.rc the entries
will need to be manually added to ExtData.rc as well.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>