Bug fix: Change time cycle flag for boundary conditions to EFY to update daily
The time cycle flag for boundary conditions was previously EY, which caused
the BC files to only be read once at the beginning of the simulation. To avoid
this, we change the time cycle flag to EFY which does not force Dta%UpdtFlag =
HCO_UFLAG_ONCE in HEMCO's hco_config_mod.F90.

This fixes the issue raised in https://github.com/geoschem/geos-chem/issues/700.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>