Now use _SSA instead of _SALA and _SSC instead of _SALC in State_Het
Some of the variable names in State_Het used the _SALA and _SALC suffixes
(e.g. Br_conc_SALA, Cl_conc_SALA).  But this can be misleading because
in the case of Br_Conc_SALA, this is Br- concentration in sea salt
aerosol which comes from BrSALA.  Therefore, to make things more
generic and to also line up with the SSA and SSC parameters for
the aerosol types, we now use _SSA and _SSC suffices for fields
of the State_Het object.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>