Add fix for CEDS emissions in tagged CO simulation
The CEDS entries in HEMCO_Config.rc.tagCO were still using the sectors from
the CEDS GBD-MAPS (NRTR, ROAD, RCOR, RCOO) but should now be using those
from the CEDS v2 (TRA, RCO).

Also fix the logicals for the CHINA_MASK in that file so that it is read if
CEDSv2, CEDS_GBDMAPS, or CEDS_GBDMAPS_byFuelType is set to true.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>