Restore KHETI_SLA(HOCl_plus_HBr) term in hetchem routine HOClUptkByHBr
We had previously commmented out the term

  k = k + H%xArea(SLA) * H%KHETI_SLA(HOCl_plus_HBr)

because this was introducing numerical noise that could not be
matched to the reference code.  The problem was due to the field
State_Chm%KHETI_SLA(9) being set to TINY(1.0e_f8) in ucx_mod.F90.
The TINY function returns the smallest possible positive value for
REAL*8 in this case, which is ~1e-308.  This value is very close to
being denormal, and can cause numerical instability.  We fixed that
issue by setting State_Chm%KHETI_SLA(9) = 0.0_f8, which avoids the
instability.

With the instability fixed, we are able to restore this term to
function HOClUptkByHBr, and we have obtained identical results
to the reference code (with the same fix in ucx_mod.F90).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>