Move init/register of State_Chm%pHcloud out of ITS_A_FULLCHEM_SIM block
The new Luo et al 2020 wet deposition scheme requires the value of
State_Chm%phCloud even for non-fullchem or aerosol routines.  We have
now moved this to the #ifdef LUO_WETDEP block where the other
Luo et al 2020-specific fields are initialized.

Also, based on a communication from Gan Luo, we make sure to set
State_Chm%pHcloud to an initial value of 5.6.  This value is appropriate
for cloud water (it accounts for dissolved CO2 in cloud water droplets).

For the fullchem and aerosol simulations, State_Chm%pHcloud will be
computed in the sulfate chemistry routines.  But for other simulations,
the default value of State_Chm%pHcloud = 5.6 will be used.

These updates are necessary to avoid an out-of-bounds error caused
by State_Chm%pHcloud not being allocated in simulations other than
fullchem or aerosol-only.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>