Bug fix: Remove extranous entries in species database for DMS
The DMS entry in the species database had a bunch of extraneous
entries following the MW_g tag.  These were left over from a Git
merge.

DMS was incorrectly tagged with Is_Aerosol=true, but did not have a
KcScale tag defined.  Thus, KcScale was set to [-999, -999, -999],
which in turn caused the exponential expression in the GET_F function to
blow up.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>