Refine algorithm in AlarmIncrementYears in History/histcontainer_mod.F90
This commit should fix the issue reported in geoschem/geos-chem #800.

The algorithm in AlarmIncrementYears (which increments the FileWriteAlarm
and FileCloseAlarm for intervals > 1 year) has been updated to:

(1) If this year is a leap year, then increment alarm by 365 days.

(2) If this year the first leap year we have encountered:
    (a) If it is after March 1st, increment alarm by 365 days
        (as we have already passed the leap year day)
    (b) Otherwise, increment alarm by 366 days.

(3) For each subsequent leap year, increment alarm by 366 days.

(4) For all other years,increment alarm by 365 days.
    then increment the alarm by 366 days.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>