Reorder diagnostics fields in HISTORY.rc.fullchem to facilitate selecting subset for benchmarking
For GEOS-Chem benchmarking, we only need to save out a subset of fields
from the DryDep, WetLossConv, and ProdLoss collections. We therefore
comment out the DryDepVel_*, Prod_* (other than O3), Loss_* (other than O3),
and WetLossConfFrac_* fields. However, the History component does not
like a comment character (#) immediately after the '.fields: ' string and
will crash. To fix this, we simply reorder the fields to list the ones
needed for benchmarking first.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>