KPP/fullchem/gckpp.kpp now links to KPP/fullchem/fullchem.kpp
For naming consistency, the gckpp.kpp file now links to fullchem.kpp,
which reflects that this is the .kpp file for the fullchem mechanism.
This is because the current KPP code uses the filename to define
the prefix for all of the Fortran generated data files.

This allows us to keep user-modified files for the "fullchem" mechanism
to have the "fullchem_" prefix, and to have all KPP-generated files
(except the symlink gckpp.kpp) to have the prefix "gckpp_".

We use the "gckpp_" prefix (i.e. GEOS-Chem-KPP) as a standard naming
convention for the sake of compilation, because the code needs to
reference KPP-built code with via USE statements.  Therefore, the
KPP-generated output files must have the same naming convention
regardless of what the name of the mechanism (e.g. fullchem, custom, etc.)
is.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>